refactor(mapgen): consolidate authoring stages
