+++ packages/mapgen-core/src/authoring/stage/index.ts
@@ -1,0 +1,8 @@
+export { createStage, deriveStageAuthoringModel } from "./create.js";
+export { assertStageId, StageIdSchema } from "./identity.js";
+export type {
+  StageAuthoringConfigLayer,
+  StageAuthoringModel,
+  StageAuthoringRuntimeStep,
+  StageObservation,
+} from "./types.js";
