+++ packages/mapgen-core/test/authoring/stage/create.test.ts
